backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
openai>=1.30.0
anthropic>=0.25.0
google-generativeai>=0.5.0
